requirements.txt
openai
langchain
pinecone-client
tiktoken
wikipedia
pypdf
docx2txt
langchain-openai
langchain-pinecone
python-dotenv
